fix: move required handling and add default to number widget
This adds a default for any number field that is not marked as required in proto. It also moves the required definition to the field generation in order to be able to check if the field is required or not. This is done to avoid setting a default for required fields as that could lead to the user missing a required field